fix(optimizer): address checkpoint/quant/WebGpu review findings on #719
Resolves the critical/major CodeRabbit findings on the fused-optimizer work:

- FusedOptimizer int8 block-quant: reject non-positive blockSize (0 spun forever,
  negative walked out of bounds); decode a zero-scale block as all-zero moments
  instead of flooring the scale to 1e-10 (which decoded default/zero quant bytes
  into spurious moments and nudged params on the first step).
- Checkpoint deserialization: bound every serialized count against the remaining
  stream before allocating (a corrupt count could OOM the process), and validate
  each LR-schedule Kind + array arity at read time so a malformed schedule fails
  loudly at load rather than IndexOutOfRange deep inside GetLr mid-training.
- LR schedules: the built-in LinearWarmup was the ONE built-in with no checkpoint
  capture, so a compiled plan using it silently lost its schedule on resume — add
  a LinearWarmupDecay kind + a bit-identical capture/restore + round-trip tests.
- WebGpu BF16 Adam/AdamW: the step field `t` was packed as Int32BitsToSingle(step)
  but the bf16 shader read it as an f32 value (pow(beta, t), t <= 1.0), so step 1
  became ~1.4e-45 → bias correction divided by zero and every step looked like the
  first. Declare `t: u32` (matching the fp32 shader) and convert with f32() in-shader.
- Compiled GPU moments: int8 V-quant and bf16 m/v buffers were AllocateByteBuffer'd
  but never zero-initialized (device malloc returns garbage) — upload zeros so the
  first step doesn't read uninitialized moments.

142 optimizer/serialization tests green (incl. new LinearWarmup round-trip + malformed
-arity rejection tests).

+++ src/AiDotNet.Tensors/Engines/DirectGpu/WebGpu/WebGpuKernels.cs
@@ -1250,7 +1250,7 @@ struct OptimizerParams {
     beta2: f32,
     epsilon: f32,
     weight_decay: f32,
-    t: f32,
+    t: u32,
 }
 @group(0) @binding(4) var<uniform> opt_params: OptimizerParams;
 
@@ -1318,7 +1318,7 @@ fn adam_bf16(@builtin(global_invocation_id) gid: vec3<u32>) {
     if (opt_params.weight_decay > 0.0) {
         grad = grad + opt_params.weight_decay * params_arr[idx];
     }
-    let first_step = opt_params.t <= 1.0;
+    let first_step = opt_params.t <= 1u;
     let old_m = select(bf16_to_float(load_m(idx)), 0.0, first_step);
     let old_v = select(bf16_to_float(load_v(idx)), 0.0, first_step);
     let m_new = opt_params.beta1 * old_m + (1.0 - opt_params.beta1) * grad;
@@ -1326,8 +1326,8 @@ fn adam_bf16(@builtin(global_invocation_id) gid: vec3<u32>) {
     store_m(idx, float_to_bf16_rne(m_new));
     store_v(idx, float_to_bf16_rne(v_new));
 
-    let m_hat = m_new / (1.0 - pow(opt_params.beta1, opt_params.t));
-    let v_hat = v_new / (1.0 - pow(opt_params.beta2, opt_params.t));
+    let m_hat = m_new / (1.0 - pow(opt_params.beta1, f32(opt_params.t)));
+    let v_hat = v_new / (1.0 - pow(opt_params.beta2, f32(opt_params.t)));
     params_arr[idx] = params_arr[idx] - opt_params.lr * m_hat / (sqrt(v_hat) + opt_params.epsilon);
 }
 
@@ -1343,7 +1343,7 @@ fn adamw_bf16(@builtin(global_invocation_id) gid: vec3<u32>) {
         params_arr[idx] = params_arr[idx] * (1.0 - opt_params.lr * opt_params.weight_decay);
     }
 
-    let first_step = opt_params.t <= 1.0;
+    let first_step = opt_params.t <= 1u;
     let old_m = select(bf16_to_float(load_m(idx)), 0.0, first_step);
     let old_v = select(bf16_to_float(load_v(idx)), 0.0, first_step);
     let m_new = opt_params.beta1 * old_m + (1.0 - opt_params.beta1) * grad;
@@ -1351,8 +1351,8 @@ fn adamw_bf16(@builtin(global_invocation_id) gid: vec3<u32>) {
     store_m(idx, float_to_bf16_rne(m_new));
     store_v(idx, float_to_bf16_rne(v_new));
 
-    let m_hat = m_new / (1.0 - pow(opt_params.beta1, opt_params.t));
-    let v_hat = v_new / (1.0 - pow(opt_params.beta2, opt_params.t));
+    let m_hat = m_new / (1.0 - pow(opt_params.beta1, f32(opt_params.t)));
+    let v_hat = v_new / (1.0 - pow(opt_params.beta2, f32(opt_params.t)));
     params_arr[idx] = params_arr[idx] - opt_params.lr * m_hat / (sqrt(v_hat) + opt_params.epsilon);
 }
 ";
